requirements.txt
livekit-agents~=0.11.0
livekit-plugins-openai~=0.10.5
livekit-plugins-silero~=0.7.3
langchain_chroma~=0.1.4
langchain_ollama~=0.2.0
pyperclip~=1.9.0